requirements.txt
AnyQt==0.0.10
attrs==19.3.0
biopython==1.74
Bottleneck==1.2.1
CacheControl==0.12.5
certifi==2019.9.11
chardet==3.0.4
commonmark==0.9.1
cycler==0.10.0
decorator==4.4.0
docutils==0.15.2
entrypoints==0.3
future==0.18.1
Genesis-PyAPI==1.2.1
GEOparse==1.2.0
idna==2.8
joblib==0.14.0
keyring==19.2.0
keyrings.alt==3.1.1
kiwisolver==1.1.0
llvmlite==0.30.0
lockfile==0.12.2
matplotlib==3.1.1
msgpack==0.6.2
networkx==2.4
numba==0.46.0
numpy==1.17.3
openTSNE==0.3.11
orange-canvas-core==0.1.7
orange-widget-base==4.0.1
Orange3==3.23.1
Orange3-Bioinformatics==4.0.0
Orange3-spark==0.2.7
pandas==0.25.2
point-annotator==1.0.0
py4j==0.10.8.1
pyclipper==1.1.0.post1
pynndescent==0.3.3
pyparsing==2.4.2
PyQt5==5.13.1
PyQt5-sip==12.7.0
pyqtgraph==0.10.0
PyQtWebEngine==5.13.1
pyswarms==1.1.0
python-dateutil==2.8.0
python-louvain==0.13
pytz==2019.3
pywin32-ctypes==0.2.0
PyYAML==5.1.2
requests==2.22.0
requests-cache==0.5.2
scikit-learn==0.21.3
scipy==1.3.1
serverfiles==0.3.0
six==1.12.0
slumber==0.7.1
sqlparse==0.3.0
tqdm==4.36.1
urllib3==1.25.6
xlrd==1.2.0
XlsxWriter==1.2.2
